+++ src/timer.h
@@ -11,6 +11,6 @@ struct emu_millis_timer{
 struct emu_millis_timer* init_timer(struct emu_millis_timer* t);
 uint64_t millis_since(struct emu_millis_timer* t);
 void reset(struct emu_millis_timer* t);
-void emu_delay(uint64_t millis);
+void emu_delay(double millis);
 
 #endif
